Code Refactoring #2

+++ src/masternodeman.cpp
@@ -591,7 +591,7 @@ CMasternode* CMasternodeMan::GetNextMasternodeInQueueForPayment(int nBlockHeight
             continue;
 
         //make sure it has as many confirmations as there are masternodes
-        if (mn.GetMasternodeInputAge() < nMnCount)
+        if (mn.GetMasternodeInputAge() < (int)nMnCount)
             continue;
 
         vecMasternodeLastPaid.push_back(std::make_pair(mn.SecondsSincePayment(), mn.vin));
